fix hprogressbar
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@1730 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ Gui/skin/skin.h
@@ -12,7 +12,7 @@ extern int skinRead( char * dname  );
 extern int skinBPRead( char * fname, txSample * bf );
 
 extern void btnModify( int event,float state );
-extern int btnGetValue( int event );
+extern float btnGetValue( int event );
 
 // ---
